Pull request #476: Aboldyrev MM-5789 asyncs
Merge in MML/infobip-mobile-messaging-android from aboldyrev-MM-5789-asyncs to master

Squashed commit of the following:

commit f88e4dbc2eb531c7b2daf9bd9dfe59faf57d1661
Merge: 5a9a74df 9e4f1ac5
Author: Alexander Boldyrev <alexander.boldyrev@infobip.com>
Date:   Wed Jan 29 11:50:59 2025 +0300

    Merge branch 'master' into aboldyrev-MM-5789-asyncs

commit 5a9a74df4df55f5f8f6699118e6d79e8d2132d0a
Author: Alexander Boldyrev <alexander.boldyrev@infobip.com>
Date:   Wed Jan 29 11:49:30 2025 +0300

    mmasync updated

commit dc826e1876ae92d2796a961070a1be7d16bf76b0
Author: Alexander Boldyrev <alexander.boldyrev@infobip.com>
Date:   Thu Dec 26 20:35:53 2024 +0300

    cleaning up

commit 2d45540b0dfb08adf6d9af5c4ffa04d753391136
Merge: 98c1b4e2 6e56ea55
Author: Alexander Boldyrev <alexander.boldyrev@infobip.com>
Date:   Thu Dec 26 18:14:34 2024 +0300

    Merge branch 'refs/heads/master' into aboldyrev-MM-5789-asyncs

commit 98c1b4e2e161b9d371fd8c8904cf967e005d3e09
Author: Alexander Boldyrev <alexander.boldyrev@infobip.com>
Date:   Thu Dec 26 18:09:32 2024 +0300

    asynctask moved to SDK level

+++ infobip-mobile-messaging-android-sdk/src/main/java/org/infobip/mobile/messaging/mobileapi/common/MRetryableTask.java
@@ -2,6 +2,7 @@ package org.infobip.mobile.messaging.mobileapi.common;
 
 import android.os.Handler;
 import android.os.Looper;
+
 import androidx.annotation.Nullable;
 
 import java.util.concurrent.Executor;
@@ -107,7 +108,7 @@ public abstract class MRetryableTask<IN, OUT> extends IMAsyncTask<IN, OUT> {
                     public void run() {
                         MRetryableTask.this.execute();
                     }
-                }, TimeUnit.SECONDS.toMillis(executionContext.attempts * executionContext.attempts * executionContext.retryPolicy.getBackoffMultiplier()));
+                }, TimeUnit.SECONDS.toMillis((long) executionContext.attempts * executionContext.attempts * executionContext.retryPolicy.getBackoffMultiplier()));
             }
 
             @Override
